+++ client/eslint.config.js
@@ -15,9 +15,9 @@ import stylistic from '@stylistic/eslint-plugin'
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
 const compat = new FlatCompat({
-    baseDirectory: __dirname,
-    recommendedConfig: js.configs.recommended,
-    allConfig: js.configs.all
+  baseDirectory: __dirname,
+  recommendedConfig: js.configs.recommended,
+  allConfig: js.configs.all
 })
 
 // eslint.config.js
@@ -54,7 +54,7 @@ export default [
       'dist/',
       'coverage/',
       'node_modules/',
-      'yarn.lock',
+      'pnpm-lock.yaml',
       'package.json',
       '/src/serviceWorker.ts',
       'eslint.config.js',