+++ client/src/test/integration/utils/placeResultsMockApi.tsx
@@ -13,7 +13,7 @@ export default function placeResultsMockApi(): void {
   // Mock place search with the query "andy warhol"
   nock(apiUrl)
     .get(
-      '/api/search/place?q=%7B%22AND%22%3A%5B%7B%22text%22%3A%22andy%22%2C%22_lang%22%3A%22en%22%7D%2C%7B%22text%22%3A%22warhol%22%2C%22_lang%22%3A%22en%22%7D%5D%7D&page=1&sort=',
+      '/api/search/place?q=%7B%22AND%22%3A%5B%7B%22text%22%3A%22andy%22%2C%22_lang%22%3A%22en%22%7D%2C%7B%22text%22%3A%22warhol%22%2C%22_lang%22%3A%22en%22%7D%5D%7D&page=1&pageLength=20',
     )
     .reply(200, JSON.stringify(mockResults(mockPlaceUri, 10)), {
       'Access-Control-Allow-Origin': '*',
@@ -22,9 +22,13 @@ export default function placeResultsMockApi(): void {
 
   // Mock the facets requests and return since they are not being tested with this mock api
   for (const facet of facetNamesLists.places) {
+    let sort = ''
+    if (facet.includes('Date')) {
+      sort = '&sort=asc'
+    }
     nock(apiUrl)
       .get(
-        `/api/facets/place?q=%7B%22AND%22%3A%5B%7B%22text%22%3A%22andy%22%2C%22_lang%22%3A%22en%22%7D%2C%7B%22text%22%3A%22warhol%22%2C%22_lang%22%3A%22en%22%7D%5D%7D&name=${facet}`,
+        `/api/facets/place?q=%7B%22AND%22%3A%5B%7B%22text%22%3A%22andy%22%2C%22_lang%22%3A%22en%22%7D%2C%7B%22text%22%3A%22warhol%22%2C%22_lang%22%3A%22en%22%7D%5D%7D&name=${facet}${sort}&page=1`,
       )
       .reply(200, JSON.stringify(null), {
         'Access-Control-Allow-Origin': '*',